Refine testimonial rating class name
The CSS class `.testimonial-star-rating` was renamed to `.testimonial-rating` to better reflect its purpose and to not limit the rating representation to stars only. This name change allows for future flexibility in the type of ratings we might want to implement, without tying the class name strictly to the star iconography.